cmake: use ENV variable BOOST_SOURCE_DIR if present
if BOOST_SOURCE_DIR is defined in ENV variable,
and is not overruled by -DBOOST_SOURCE_DIR at cmake invocation
then we use it.

If BOOST_SOURCE_DIR is not defined in ENV variable,
and not defined by -DBOOST_SOURCE_DIR,
then we look for the system compiled boost.

-> allows to set up BOOST_SOURCE_DIR in your bashrc
on machine where boost compilation is needed, once for all