Fix #8
- Order of inclusion of Python include path and nunpy python include path
reversed.
--> on Ubuntu 12.04, there is a link numpy in /usr/include/python2.7
which was *before* the numpy include path, i.e.  in this case
the new version of numpy installed by virtualenv.

By reversing the order of the include, the numpy arrayobject.h
file of virtualenv is found first, the Python.h is found anyway
in the correct system directory (does not change with new numpy version).

O. Parcollet