New version of time_pt
The time points have been changed. Now there are two layers:

- a time segment
- time points that are generated from the time segment

Once a segment is defined it is not necessary to pass beta as an argument. All
points are generated for the same interval. This avoids some confusions.

A new convention is chosen: the point at Nmax == beta == 0. This point can
never be obtained by additions or multiplications. So the largest interval has
Nmax-1. This is good because when a random number is generated with rand(Nmax)
it takes values between 0 and Nmax-1 which is the largest interval.