[gfs] New indices gf_indices
This commit is replacing indices_2 by gf_indices_pair, a pair of gf_indices.
The gf_indices are essentially a vector<string>.  There is a converter so that
gf_indices can be converted both from a python sequence of strings or integers.

I also fixed the slicing of indices in this commit.