[det] Fix a sign in check_mat_inv
Warning: in the det_manip class the sign is not necessarily the sign of the
rearrangement of the rows and cols. See in particular try_remove. There, the
sign of the det comes from the displacement of a row and a col to the end of
the matrix, namely (i+j)%2. However, in the repacking of the matrix we
*exchange* the last row/column with i and j. These two swaps would not
lead to any change in the sign (1+1)%2=0.