[gfs] fit_tail should be working now
There was a very misleading name for one of the arguments of fit_tail.
n_moments was supposed to be both the highest moment to be fitted and the
number of moments that are fitted. Those two things do not need to be the same
(unless order_min in the known tail is 1). Therefore, I renamed the variable to
max_moment and changed the code in order to have the following interpretation:

max_moment: the highest moment to be fitted and therefore also the highest
meaningful moment in the tail after the fit.