[tails] Fix #67: conjugate of a tail
The tail of a conjugate Green's function is not in general the
conjugate of the tail. It depends wether the Green's function
is imaginary (imaginary time, imaginary frequencies) or not.
The default conj() just takes the conjugate of the tail. I
specialize it in imfreq.hpp and imtime.hpp to have a slightly
different behavior there.