[gfs] fourier tfm: tensr-v'ed and two-variable gfs
Tensor-valued GF with possibility of using a tail (contrary to the 'native' TRIQS FT of tensor-valued functions, which does not take into account the tail since tensor-valued functions do not have a tail until now).
Two-variable FT: a bit wordy, not using curry because would have to call 'native' TRIQS FT, which does not have a tail.

Remark: would be nice to have a tail for gf<imfreq, tensor_valued>, i.e a tail with tensor-valued coefficients