[test] Remove serialization test by default for variant.
By default, the test is not done any more.
It requires boost libs, and it is the only test in the lib that requires
it. This leads to compilation mess with gcc > 5.2, icc depending
on the version of gcc that compiled boost.

Can be put back with a macro if needed.