[gf] Fix #288
The tail_zero was not properly reloaded from h5.
tail_zero has a state (the size of the matrix basically).
To avoid writing a matrix full of 0 in the h5 file,
I split the writing/reading of the singularity in a small trait,
and specialize it for tail_zero : writing does nothing,
reading resizes the tail accordingly to the target shape.