[det_manip] Fix overflow issue (partially).
If the det becomes very small, it can under/overflow
even when the det *ratios* are finite.
Change the implementation so that we only work with ratios,
we never use the det.

The det is still stored for optional usage by the user,
but is not used to compute the try_xxx ratios any more.

In case of overflow, it is the responsability of the user of the class
to act on it. Normally the try_xxx ratios will not over/underflow.

Author:    JaksaVucicevic <jaksa.vucicevic@gmail.com>

Review by O. Parcollet.