[cmake] Fix #292
The add_definition in TRIQSConfig.cmake is an error.
It adds definitions to all compilers, including fortran, leading to compilation error.
The fix is to remove it.

HOWEVER, it means that it must now be added in all applications
explicitely.

TRIQSConfig should only set variables, and define macros,
it should not add_definitions, add_anything behind the scene.
It is now the case after this fix.

ChangeLog
-----------

[Applications CMakeLists]
Due to a change in the TRIQSConfig.cmake, necessary for proper compilation
of f90 codes in dft_tools on OS X (issue https://github.com/TRIQS/dft_tools/issues/19),
one now needs to explicitely add the TRIQS_CXX_DEFINITIONS in the cmakelist.

Add to main CMakeList of your application (after find_package(TRIQS))

add_definitions(${TRIQS_CXX_DEFINITIONS})