[cmake] Solve the infamous "make install before test" problem.
Problem
-------

When one runs a test of a TRIQS-based application 'myapp', dynamic linker
looks for libmyapp_c.(so|dylib) first in ${TRIQS_PATH}/lib, and only after
that in the build directory of myapp. Version mismatch between the installed
shared library and test executables usually leads to a crash. This is
why we have to precede a call to `make test` with a call to `make install`
if libmyapp_c.(so|dylib) already exists in the installation dir.

Solution
--------

* Create a new IMPORTED library target in TRIQSConfig.cmake. This
  target is called 'triqs', and it points at the *installed*
  libtriqs.(so|dylib).

* TRIQS_LIBRARY var is now set to 'triqs' in TRIQSConfig.cmake,
  i.e. it refers to the imported library target. TRIQS_LIBRARY_ALL
  is updated accordingly.

* triqs_set_rpath_for_target() macro does not set RPATH linker flags
  explicitly.

* With this set of changes CMake links application executables and
  libs such that their RPATH begins with directories in the *build*
  tree. These parts of RPATH are removed upon installation.

* Additionally sanitize test environment by unsetting LD_LIBRARY_PATH
  and DYLD_LIBRARY_PATH in triqs_add_cpp_test() and triqs_add_python_test()
  macros.