Fix #441
Change the index type of matsubara mesh to _long,
i.e. a long in a struct, forbidding any operations.
matsubara_freq can now unambigously be casted in :
  - complex for any operation, as before
  - _long, just for indexing.

As a result, the g[x] and g[{x,y,z}]
can be called with matsubara_freq now as well as long and mesh_point.
This makes the code compile properly.