Disable automatic tailfit in fourier transform of real-frequency Green functions
The usage of high-frequency moments in the real-frequency fourier
transform is much more delicate compared to the Matsubara variant.
In order to guarantee that the forward and backward transform cancel
each other proper high frequency moments need to be given in both
directions. This is tied to the fact that the number of frequency
and time mesh-points is currently enforced to be identical.
Further, some Green functions on the real axis are not analytical
for w->infty (e.g. lesser Gf), so a high-frequency expansion is
not even viable.