[sumk] assert dimensions of Sigma to match hopping
Add a check and extend the doc strings to make sure that each block in
Sigma has the same dimension as self.hopping. Otherwise sumk can give wrong
results without throwing an error. later on when [tmp - mu_pat] is
calculated the dimension will always be casted to match, because of numpy subtraction.

Thanks @phibeck for reporting the problem.