changed TB_to_sympy method to handle only 2D systems
This is because, in the unit tests, the set up involves a simple 2D system. Instead of changing the unit tests to handle a simple 3D system, I changed my function to handle a 2D system. This did not require a lot of work on my end. Perhaps at a later date, we can incorporate tests for 3D systems.